// src/app/@core/enums/school-types.ts
export enum SchoolTypes {
	GUINEA_SCHOOL = "GUINEA_SCHOOL",
	GUINEA_PRIMARY = "GUINEA_PRIMARY",
	GUINEA_SECONDARY = "GUINEA_SECONDARY",
	TANZANIA_PRIMARY = "TANZANIA_PRIMARY",
	TANZANIA_SECONDARY = "TANZANIA_SECONDARY",
	KCPE_PRIMARY_SCHOOL = "KCPE_PRIMARY_SCHOOL",
	KCSE_SCHOOL = "KCSE_SCHOOL",
	IGCSE_SCHOOL = "IGCSE_SCHOOL",
	OLEVEL_SCHOOL = "OLEVEL_SCHOOL",
	IVORY_COAST_SCHOOL = "IVORY_COAST_SCHOOL",
	IVORY_COAST_PRIMARY = "IVORY_COAST_PRIMARY",
	IVORY_COAST_SECONDARY = "IVORY_COAST_SECONDARY",
	ZIMBABWE_PRIMARY = "ZIMBABWE_PRIMARY",
	ZIMBABWE_SECONDARY = "ZIMBABWE_SECONDARY",
	ZIMBABWE_IGCSE = "ZIMBABWE_IGCSE",
	SOUTH_AFRICA_PRIMARY = "SOUTH_AFRICA_PRIMARY",
	SOUTH_AFRICA_SECONDARY = "SOUTH_AFRICA_SECONDARY",
	ZAMBIA_PRIMARY = "ZAMBIA_PRIMARY",
	ZAMBIA_SECONDARY = "ZAMBIA_SECONDARY",
	GHANA_PRIMARY = "GHANA_PRIMARY",
	GHANA_PRIMARY_JUNIOR = "GHANA_PRIMARY_JUNIOR",
	GHANA_JUNIOR = "GHANA_JUNIOR",
	GHANA_SENIOR = "GHANA_SENIOR",
}